+++ Projection.hs
@@ -5,15 +5,15 @@ import Vec
 import Geometry (Ray(..))
 
 {- Projection -}
-data Projection = Orthographic { width :: Float, height :: Float}
-                | Perspective { fovy :: Float
-                              , width :: Float
-                              , height :: Float
-                              , near :: Float }
+data Projection = Orthographic { width :: Double, height :: Double}
+                | Perspective { fovy :: Double
+                              , width :: Double
+                              , height :: Double
+                              , near :: Double }
                   deriving Show
 
 {- Create ray from camera givent the projection -}
-rayFromPixel :: Float -> Float -> Projection -> Float -> Float -> Ray
+rayFromPixel :: Double -> Double -> Projection -> Double -> Double -> Ray
 rayFromPixel w h (Orthographic pw ph) x y = Ray o (Vec 0 0 1) where
   o = Vec (apw * (x - (w/2)) / w) (aph * ((-y) + (h/2)) / h) 0
   aspect = w / h
