+++ Vec.hs
@@ -16,7 +16,7 @@ module Vec (Vec(Vec)
            , fromRGB) where
 import Color hiding (mul)
 
-data Vec = Vec Float Float Float deriving (Eq, Show)
+data Vec = Vec Double Double Double deriving (Eq, Show)
 instance Num Vec where
   (Vec x1 y1 z1) + (Vec x2 y2 z2) = (Vec (x1+x2) (y1+y2) (z1+z2))
   (Vec x1 y1 z1) * (Vec x2 y2 z2) = (Vec (x1*x2) (y1*y2) (z1*z2))
@@ -35,23 +35,23 @@ yAxis = Vec 0 1 0
 zAxis :: Vec
 zAxis = Vec 0 0 1
 
-mul :: Float -> Vec -> Vec
+mul :: Double -> Vec -> Vec
 mul l (Vec x y z) = Vec (l*x) (l*y) (l*z)
 
-dot :: Vec -> Vec -> Float
+dot :: Vec -> Vec -> Double
 dot (Vec x1 y1 z1) (Vec x2 y2 z2) = x1*x2 + y1*y2 + z1*z2
 
 cross :: Vec -> Vec -> Vec
 cross (Vec x1 y1 z1) (Vec x2 y2 z2) =
   Vec (y1*z2 - z1*y2) (z1*x2 - x1*z2) (x1*y2 - y1*x2)
 
-sqrLen :: Vec -> Float
+sqrLen :: Vec -> Double
 sqrLen v = dot v v
 
-sqrDist :: Vec -> Vec -> Float
+sqrDist :: Vec -> Vec -> Double
 sqrDist v w = sqrLen (v - w)
 
-dist :: Vec -> Vec -> Float
+dist :: Vec -> Vec -> Double
 dist v w = sqrt (sqrDist  v w)
 
 normalize :: Vec -> Vec
@@ -66,7 +66,7 @@ fromRGB (RGB x y z) = (Vec x y z)
 reflect :: Vec -> Vec -> Vec
 reflect v n = v - (Vec.mul (2 * (dot v n)) n)
 
-refract :: Vec -> Vec -> Float -> Float -> Maybe Vec
+refract :: Vec -> Vec -> Double -> Double -> Maybe Vec
 refract i n n1 n2
   | sin2θ > 1 = Nothing
   | otherwise = Just $ (Vec.mul n1n2 i)
